parser: free unused token strings in token_accept()
If token_accept() is provided a null token pointer argument, the
previous "current" token is discarded.  If the token matched is a
symbol type, the token string (which will have been dynamically
allocated in yylex()) will be leaked.

Fix this by freeing the current token string if token pointer passed
is null.

The compiler warns when we attempt to free a pointer to constant
data, so change the type of the string pointer in the symbol
structure to be pointer to non-constant.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-5-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>