parser: free unused type token strings
In qmi_message_parse(), the type token for each field parsed will
contain a dynamically-allocated string continaing the field type.
We only use the type identifier (number) though, and as a result
the type name gets leaked.  Fix this by freeing the type token
string if non-null.

Essentially the same thing occurs in qmi_struct_parse() while
parsing the fields of a message.  Fix the problem there as well.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-6-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>