parser: use CHAR_MAX to define non-ASCII token id values
The first token id is given the value 256.  This ensures it is
outside the range of any valid character.

But we don't allow non-ASCII characters on input, so we can use 128
as the first non-character token id value.  Represent this as
(CHAR_MAX + 1), to make explicit the reason this value is chosen.

Give the enumerated type representing token ids a name.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-12-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>