parser: introduce memalloc() macro
Create a new macro that allocates and zeroes a block of memory,
which guarantees that the allocation will succeed.  Use this in two
spots where calloc() is already assumed not to fail, and use it in
other places where memory is dynamically allocated.

If the malloc() call in the macro fails, memalloc() will call errx()
to print an error message to stderr and exit with status 1.

In addition, check for a null pointer returned by strdup() in
yylex(), and report a similar error if that occurs.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-17-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>