parser: use stronger typing in the symbol structure
Refine the definition of the symbol structure with an anonymous
union to make it a little clearer whether a token describes a
message or a type.

In symbol_add(), be more explicit about the type of additional
arguments expected to be seen.  Add a name to the enumerated type
representing the defined "type" type values (U8, STRUCT, etc.).
And change the type of the type field in the symbol structure to
have that enumerated type.

Specifically, if the symbol being added is a message, the argument
that follows should be one of the message types (request, response,
or indication).  And if the symbol being added is a type, the next
argument is expected to be one of the "type" types.  This makes it a
little easier to understand what the code is doing.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-20-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>