parser: reset fixed flag each for each array
If a message contains an array, we need to reset the flag that
indicates whether it is a fixed array or not each time through the
loop parsing message members.  Otherwise a non-fixed array declared
after a fixed array will be marked as fixed.

Drop the "int" in the definition of the array_size local varaible in
qmi_message_parse() to be consistent with the rest of the program.

Signed-off-by: Alex Elder <elder@linaro.org>
Message-Id: <20211001232338.769309-22-elder@linaro.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>