parser: don't treat 8 as a valid octal digit
The function isodigit() was defined to mimic isxdigit(), indicating
whether a given character was an octal digit.  But as written, it
considers '8' to be a valid octal digit, which it is not.  Fix this
bug.

In addition, the "bad_octal.qmi" test file refers to an undefined
"test_struct" as a message member type.  Fix that by using u32
instead.

Fixes: 61f6fe9 ("parser: be more restrictive when parsing numbers")
Signed-off-by: Alex Elder <elder@linaro.org>